Feature #372: Remove appassembler dependency.
The use of appassembler was eliminated by constructing several Ant build scripts.
build.xml contains the target apps which shall be built and their parameters. The parameters are the same as for the appassembler approach.
build_*_script.xml are used to generate the startup scripts of the targets.
build.xml is invoked by the pom.xml in the sdk-package directory.