Feature #376: Introduce logger.properties to NMF.
Each log uses the logging.properties file in the root directory of this repository. This file can be placed inside an arbitrary directory. However, the environment variable NMF_HOME has to point to that directory. Every logger generates its own logfile for which the log levels can be changed inside logging.properties. The console output stays reserved for log entries of level WARNING or above.