Feature #1061 Merge Mission Planning back to dev.
The implementation of the Mission Planning (MP) services includes
the following services:

* Planning Request Service allows the submission of Planning Requests
  with their associated status tracking.

* Plan Distribution Service provides access to plans and allows
  monitoring plans and plan statuses.

* Plan Information Management Service provides operations for managing
  the Planning Request, planning activity, planning event and planning
  Resource Definitions.

* Plan Edit Service provides the operations for editing the currently
  executing Plan (or Schedule). Operations are provided to Insert,
  Update and Delete occurrences of Planning Activities and Events.

* Plan execution control service provides the capability for submitting
  plans.

Additionally a Platform.Clock service was added.
The motivation for such service came from MP Plan Execution Control
Service implementation, where activities are polled for execution
every second. In order for this to work with Simulator, where time
can be shifted and accelerated, the system time needs to be provided
by the simulator. Thus a Clock service is defined in
ServiceDefPLATFORM.xml with methods "getTime" and "getTimeFactor".
The Clock is implemented in esa.mo.helpertools.clock.SystemClock.

COM changes:
  TransactionsProcessor::insert - fix to allow retry on failed insert
  to COM Archive insert queue.